Temp dir test fixture now implicitly available across all modules
